splot plotting gridlines into separate fucntion
